수정중

+++ Spring/great/src/main/java/com/kh/great/domain/Product.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Product {
-    private Integer productNumber;      //상품번호  P_NUMBER	VARCHAR2(4 BYTE)	No
+    private Long productNumber;      //상품번호  P_NUMBER	VARCHAR2(4 BYTE)	No
     private Integer owenerNumber;       //점주번호  OWNER_NUMBER	NUMBER(6,0)	No
     private String storeName;           //가게명   STORE_NAME	VARCHAR2(30 BYTE)	No
     private String title;               //글제목   P_TITLE	VARCHAR2(300 BYTE)	No
